% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{compare}
\alias{compare}
\title{Generate simulation data and evaluate estimators, with sample splitting.}
\usage{
compare(n, ex_setting, f_setting, nfold = 5)
}
\arguments{
\item{n}{integer number of samples. For "401k" ex_setting this is ignored
and the whole data set is used.}

\item{ex_setting}{string "normal", "mixture2", "mixture3",
"logistic", "t4", "401k".}

\item{f_setting}{string "plm", "additive", "interaction".}

\item{nfold}{integer number of cross-validation folds.}
}
\value{
list containing estimates, standard error estimates, and sample theta (for debugging).
}
\description{
Generate simulation data and evaluate estimators, with sample splitting.
}
