% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{compare_evaluate}
\alias{compare_evaluate}
\title{Evaluate estimators by training nuisance functions on training set and evaluating them on test set.}
\usage{
compare_evaluate(train, test, ex_setting, f_setting, regression, sm_bw_out)
}
\arguments{
\item{train}{list containing vector of responses y and matrix of predictors X = (x,z).}

\item{test}{list containing vector of responses y and matrix of predictors X = (x,z).}

\item{ex_setting}{string "normal", "mixture2", "mixture3",
"logistic", "t4", "401k".}

\item{f_setting}{string "plm", "additive", "interaction".}

\item{regression}{Optional fitted regression.}

\item{sm_bw_out}{Output of cv_resmooth.}
}
\value{
list containing f, df, and score estimates evaluated on the test set.
}
\description{
Evaluate estimators by training nuisance functions on training set and evaluating them on test set.
}
