% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{compare_rothenhausler}
\alias{compare_rothenhausler}
\title{Generate simulation data and evaluate Rothenhausler estimator.}
\usage{
compare_rothenhausler(n, ex_setting, f_setting)
}
\arguments{
\item{n}{integer number of samples. For "401k" ex_setting this is ignored
and the whole data set is used.}

\item{ex_setting}{string "normal", "mixture2", "mixture3",
"logistic", "t4", "401k".}

\item{f_setting}{string "plm", "additive", "interaction".}
}
\value{
list containing estimate, standard error estimate, and sample theta (for debugging).
}
\description{
Generate simulation data and evaluate Rothenhausler estimator.
}
