% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{fit_lasso_poly}
\alias{fit_lasso_poly}
\title{Fit a lasso regression using quadratic polynomial basis, with interactions.}
\usage{
fit_lasso_poly(X, y, degree, lambda = NULL)
}
\arguments{
\item{X}{matrix of covariates.}

\item{y}{vector of responses.}

\item{degree}{maximum degree of polynomial terms.}

\item{lambda}{optional scalar tuning parameter, if "NULL" chosen via
cross-validation.}
}
\value{
List containing: A function "fit" which takes matrix input of the
same width as X, and returns a vector of y-predictions. A
scalar "lambda" the tuning parameter.
}
\description{
Compute regression function and derivative estimates based
on polynomial basis lasso with penalty parameter chosen by
cross validation (CV).
}
