% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{fit_xgboost}
\alias{fit_xgboost}
\title{Fit pre-tuned XGBoost regression for use in simulations.}
\usage{
fit_xgboost(X, y, params, derivative = FALSE)
}
\arguments{
\item{X}{matrix of covariates.}

\item{y}{vector of responses.}

\item{params}{XGBoost hyperparameters.}

\item{derivative}{logical determining if numerical difference derivative
estimate (wrt the first predictor) should also be returned.}
}
\value{
list containing a function "fit" which takes matrix input of the
same width as X, and returns a vector of predictions. Optionally the list
also contains a function "deriv_fit" for numerical difference derivative
estimates.
}
\description{
Fit pre-tuned XGBoost regression for use in simulations.
}
