% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmooth.R
\name{MC_sums}
\alias{MC_sums}
\title{Compute sums of a Monte Carlo vector for use in resmoothing.}
\usage{
MC_sums(a, n, nMC, nbw)
}
\arguments{
\item{a}{vector of length (n x nMC x nbw).}

\item{n}{integer.}

\item{nMC}{integer.}

\item{nbw}{integer.}
}
\value{
list with nbw elements. The j'th element of which is
a vector of length n, the i'th element being the sum of the
(((j-1)n + (i-1)) x nMC + 1) to (((j-1)n + i) x nMC)
elements of a inclusive.
}
\description{
Compute sums of a Monte Carlo vector for use in resmoothing.
}
