% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{mixture_score}
\alias{mixture_score}
\title{Population score function for the symmetric mixture two Gaussian random variables.}
\usage{
mixture_score(x, sd)
}
\arguments{
\item{x}{vector of observations.}

\item{sd}{standard deviation of each Gaussian.}
}
\value{
vector of length n
}
\description{
Population score function for the symmetric mixture two Gaussian random variables.
}
