% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmooth.R
\name{new_X}
\alias{new_X}
\title{Generate a matrix of covariates for use in resmoothing, in which
the d'th column of X is translated successively by the Kronecker
product of bw and MC_variates.}
\usage{
new_X(X, d, MC_variates, bw)
}
\arguments{
\item{X}{matrix of covariates.}

\item{d}{integer index of covariate to be smoothed along.}

\item{MC_variates}{vector of standard Gaussian rvs.}

\item{bw}{vector of bandwidths for the Gaussian kernel.}
}
\value{
matrix with ncol(X) columns and (nrow(X)length(MC_variates)
length(bw)) rows.
}
\description{
Generate a matrix of covariates for use in resmoothing, in which
the d'th column of X is translated successively by the Kronecker
product of bw and MC_variates.
}
