% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{ng_pseudo_response}
\alias{ng_pseudo_response}
\title{Generate pseudo responses as in Ng 1994 to enable univariate score
estimation by standard smoothing spline regression.}
\usage{
ng_pseudo_response(x, w = rep(1, length(x)))
}
\arguments{
\item{x}{vector of covariates.}

\item{w}{vector of weights.}
}
\value{
A vector of score estimates.
}
\description{
Pseudo responses should be regarded as a computational tool, not
as an estimate of the score itself.
}
\examples{
x <- seq(-3,3, length.out=50)
ng_pseudo_response(x)
}
