% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{partially_linear}
\alias{partially_linear}
\title{Fit a doubly-robust partially linear regression using the DoubleML package
and pre-tuned XGBoost regressions, for use in simulations.}
\usage{
partially_linear(X, y, g_params, m_params)
}
\arguments{
\item{X}{matrix of covariates.}

\item{y}{vector of responses.}

\item{g_params}{XGBoost hyperparameters for partially linear regression of y on X.}

\item{m_params}{XGBoost hyperparameters for predictor regression of the first column of X on the others.}
}
\value{
List containing the linear parameter estimate and the
corresponding standard error estimate.
}
\description{
Fit a doubly-robust partially linear regression using the DoubleML package
and pre-tuned XGBoost regressions, for use in simulations.
}
