% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{rmixture}
\alias{rmixture}
\title{Symmetric mixture two Gaussian random variables.}
\usage{
rmixture(n, sd)
}
\arguments{
\item{n}{number of observations.}

\item{sd}{standard deviation of each Gaussian.}
}
\value{
vector of length n
}
\description{
The resulting distribution is mean zero, variance one.
X ~ N(-sqrt(1-sd^2),sd^2) wp 0.5,
N(sqrt(1-sd^2),sd^2) wp 0.5.
}
