% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{rothenhausler_basis}
\alias{rothenhausler_basis}
\title{Generate the modified quadratic basis of Rothenhausler and Yu.}
\usage{
rothenhausler_basis(X)
}
\arguments{
\item{X}{matrix of covariates.}
}
\value{
List containing the modified basis matrices for regression and derivative estimation.
}
\description{
Generate the modified quadratic basis of Rothenhausler and Yu.
}
