% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{z_correlated_normal}
\alias{z_correlated_normal}
\title{Generate n copies of Z ~ N_{p}(0,Sigma), where Sigma_{jj} = 1, Sigma_{jk} = corr for all j not equal to k.}
\usage{
z_correlated_normal(n, p, corr)
}
\arguments{
\item{n}{integer number of samples.}

\item{p}{integer number of dimensions.}

\item{corr}{float correlation in (-1,1).}
}
\value{
n by p matrix.
}
\description{
Generate n copies of Z ~ N_{p}(0,Sigma), where Sigma_{jj} = 1, Sigma_{jk} = corr for all j not equal to k.
}
